fix GCC warnings with -flto
This issue is a compiler+linker bug when using the GNU toolchain.
The options such as `-Wno-lto-type-mismatch' and `-Wno-clobbered'
aren't passed to the linker plugin with `-flto'. Silence the
warnings.

Note that using `-flto' on Windows most likely won't result in a
working binary, either due to unresolved symbols or
miscompilation. Passing `-flto-partition=none' may alleviate some
issues but it's still likely to crash in an unexpected code path.
User reported "screenshotjpg" to break despite proper toolchain
usage.

It's also likely that `-fipa-pta' will result in an ICE at link
time regardless of the platform used (PE/COFF, ELF, or
otherwise).

The `-flto' exercise is largely a curiosity and I'd advise
against using this flag or any other advanced interprocedural
optimizations. The linker plugin is too buggy.